Minor fixes & cleanup
