--- Assets/Models/Environment/Props/Jerrycan.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 116af43efbfe5bd418ad0cc212da233d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
