+++ Assets/Prefabs/Environment/Props/Sign.prefab.meta
