--- Assets/Prefabs/Environment/Signs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1988d910a8bde454f91f1f004f196af1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
